Allow get_local_time to re-raise Exceptions after retrying
If the WiFi can't connect after several retries, it's possible
that the connection settings are misconfigured, or the signal
is too low.

In the get_local_time and _connect_esp methods, introduce an optional
reraise_exceptions flag (defaults to False), and the number of retries
before re-raising the exception (defaults to 3).

This allows the caller to then catch the raised Exception and take
action on it, such as displaying a "can't connect to wifi screen".